Refactor TodoFlow and UserFlow to use Long for ID parameters; add Mail utilities and template configuration
- Updated TodoFlow methods to accept Long instead of int for event and todo IDs.
- Modified UserFlow to include a method for retrieving user details by ID and changed addUser method to accept password instead of email.
- Introduced HionTemplateConfig for Thymeleaf template management.
- Added MailSenderFrom class for email sender configuration.
- Implemented MailUtil for sending emails asynchronously and handling password reset links.
- Enhanced AuthController with forgot-password and reset-password endpoints.
- Updated CarpoolController and EventController to use Long for ID parameters and added new endpoints for carpool and event management.
- Adjusted UserController to include user detail retrieval and modified registration to accept password.

+++ src/main/java/com/hionstudios/mypersonalinvite/controller/TodoController.java
@@ -21,7 +21,7 @@ public class TodoController {
     @PostMapping("event/{id}/add")
     @IsUser
     public ResponseEntity<MapResponse> addTodo(
-            @PathVariable int event_id,
+            @PathVariable Long event_id,
             @RequestParam String task) {
         return ((DbTransaction) () -> new TodoFlow().addTodo(event_id, task)).write();
     }
@@ -29,7 +29,7 @@ public class TodoController {
     @PutMapping("{id}/edit")
     @IsUser
     public ResponseEntity<MapResponse> updateTodo(
-            @PathVariable int id,
+            @PathVariable Long id,
             @RequestParam(required = false) String task,
             @RequestParam(required = false) Boolean status) {
         return ((DbTransaction) () -> new TodoFlow().updateTodo(id, task, status)).write();
@@ -37,13 +37,13 @@ public class TodoController {
 
     @DeleteMapping("{id}/delete")
     @IsUser
-    public ResponseEntity<MapResponse> deleteTodo(@PathVariable int id) {
+    public ResponseEntity<MapResponse> deleteTodo(@PathVariable Long id) {
         return ((DbTransaction) () -> new TodoFlow().deleteTodo(id)).write();
     }
 
     @GetMapping("event/{id}")
     @IsUser
-    public ResponseEntity<MapResponse> listTodos(@PathVariable int event_id) {
+    public ResponseEntity<MapResponse> listTodos(@PathVariable Long event_id) {
         return ((DbTransaction) () -> new TodoFlow().listTodos(event_id)).read();
     }
 }
